backend/requirements.txt
torch==2.9.1
torchvision=0.24.1
flask==3.0.0
flask-cors==4.0.0
pillow==10.1.0